when retrieving data from an inbuffer, no longer restrict ourselves to uint32

+++ include/amqpcpp/inbuffer.h
@@ -128,7 +128,7 @@ public:
      *  @param  size
      *  @return char*
      */
-    const char *nextData(uint32_t size);
+    const char *nextData(size_t size);
 
     /**
      *  The checker may access private data
